refactor: split Validator methods in separate files
