+++ packages/openapi-parser/src/lib/Validator/checkReferences.ts
@@ -1,0 +1,20 @@
+import type { Filesystem } from '../../types'
+import { resolveReferences } from './resolveReferences'
+import { transformErrors } from './transformErrors'
+
+export function checkReferences(filesystem: Filesystem) {
+  const entrypoint = filesystem.find((file) => file.entrypoint === true)
+
+  try {
+    resolveReferences(filesystem, false)
+
+    return {
+      valid: true,
+    }
+  } catch (err) {
+    return {
+      valid: false,
+      errors: transformErrors(entrypoint, err.message),
+    }
+  }
+}
