+++ packages/openapi-parser/src/lib/Validator/escapeJsonPointer.ts
@@ -1,0 +1,3 @@
+export function escapeJsonPointer(str: string) {
+  return str.replace(/~/g, '~0').replace(/\//g, '~1')
+}
