+++ packages/openapi-parser/src/lib/Validator/isObject.ts
@@ -1,0 +1,1 @@
+export const isObject = (obj: any) => typeof obj === 'object' && obj !== null
