+++ packages/openapi-parser/src/lib/Validator/unescapeJsonPointer.ts
@@ -1,0 +1,3 @@
+export function unescapeJsonPointer(str: string) {
+  return str.replace(/~1/g, '/').replace(/~0/g, '~')
+}
